+++ src/components/Product.jsx
@@ -2,15 +2,21 @@ import { useLocation } from "react-router-dom";
 import { CartContext } from "./Cart";
 import { useContext, useState } from "react";
 import InputQuantity from "./InputQuantity";
+import formatPrice from "../utilities/formatPrice";
 
 const Product = () => {
   const { state } = useLocation();
   const { product } = state;
   const { addToCart } = useContext(CartContext);
   const [quantity, setQuantity] = useState(1);
-  const addToCartHandler = (e) => {
+  const [quantityValidity, setQuantityValidity] = useState(true);
+
+  const onSubmitHandler = (e) => {
     e.preventDefault();
-    addToCart(product, quantity);
+    // const input = document.querySelector(".quantity-input");
+    // const isQuantityValid = validateInput(input);
+    const isQuantityValid = validateQuantity(quantity, setQuantityValidity);
+    isQuantityValid && addToCart(product, quantity);
   };
 
   const onChangeQuantityHandler = (newQuantity) => {
@@ -18,13 +24,17 @@ const Product = () => {
   };
 
   const incrementQuantityHandler = () => {
-    setQuantity((prevQuantity) => prevQuantity + 1);
+    setQuantity((prevQuantity) => {
+      const newQuantity = +prevQuantity + 1;
+      !quantityValidity && validateQuantity(newQuantity, setQuantityValidity);
+      return newQuantity <= 999 ? newQuantity : prevQuantity;
+    });
   };
 
   const decrementQuantityHandler = () => {
     setQuantity((prevQuantity) => {
       const newQuantity = prevQuantity - 1;
-      return newQuantity === 0 ? prevQuantity : newQuantity;
+      return newQuantity <= 0 ? prevQuantity : newQuantity;
     });
   };
 
@@ -39,9 +49,9 @@ const Product = () => {
           alt="#"
         />
       </picture>
-      <form noValidate={true} onSubmit={addToCartHandler}>
+      <form noValidate={true} onSubmit={onSubmitHandler}>
         <p>{product.description}</p>
-        <p>${product.price}</p>
+        <p>{formatPrice(product.price)}</p>
         <InputQuantity
           quantity={quantity}
           setQuantity={onChangeQuantityHandler}
@@ -52,6 +62,29 @@ const Product = () => {
       </form>
     </article>
   );
+};
+
+const onInputValidate = (e) => {
+  validateQuantity(e.target.value);
+};
+
+const validateQuantity = (quantity, callBack) => {
+  const isQuantityValid = /^(?!^0)\d{1,3}$/.test(quantity);
+  const input = document.querySelector(".quantity-input");
+  const errorMessage = document.querySelector(".error-message");
+  console.log("validateQuantity firing!");
+  if (!isQuantityValid) {
+    errorMessage.textContent =
+      "Do not leave blank. Enter a number between 1 and 999.";
+    input.addEventListener("input", onInputValidate);
+    callBack && callBack(false);
+  } else {
+    errorMessage.textContent = "";
+    input.removeEventListener("input", onInputValidate);
+    callBack && callBack(true);
+  }
+
+  return isQuantityValid;
 };
 
 export default Product;