+++ src/utilities/parseProducts.js
@@ -25,6 +25,7 @@ const parseProducts = (data) => {
     { all: [...data] }
   );
 
+  console.log(products);
   return products;
 };